Retry cache clearing when it fails
Under some circumstances (see #13) the file that is created in the document root is not directly callable. This is why I have added a retry option that could be used to retry the call if the first one failed.

Additionally one could add a retry timeout (in milliseconds) as option to the command. This way the process will wait the specified time untill the next try is started.

As soon as the file could be called, the retry-loop breaks and the process goes on as usual.